Remove NGINX Docker
In IOI 2020, we will deploy using Ansible.
We also have a dedicated Nginx host to handle routing.

There is no need to deploy an Nginx docker container.

In a past commit, the system relies on Docker to copy share the
ioi_static and ioi_media volume with the nginx container.
This has also been removed. Instead, the Translation App container
will serve the static ioi_media.