Optimize the number of editor reloads.
Currently, the editor reloads the server state whenever it refreshes
edit token.

Since the backend always issues new tokens or new sessions now, the
frontend can assume that it has the latest state if the edit token
doesn't change after refresh.

Furthermore, when the token does change, frontend can check if the
translation state is the same as server's state, and only reload the
editor when necessary.

This fixes the jumping editor issue that's caused by editor reloading.