Use num_of_contestants instead of online field
Due to various reasons, there might be hybrid teams in IOI 2022
(some contestants are on-site, some are online). It does not make sense to add
online field at the user (team) level. Fortunately we already have num_of_contestants field.
Currently, online is equivalent to num_of_contestants == 0,
so we can use it and drop the online field completely.