Reform handling of tokens in the editor
AJAX requests from the beforeunload event handler do not work with
current browsers – the page is killed before the request completes.

The beforeunload handler now asks "are you sure?" if there are unsaved data.

The unload handler uses the Beacon mechanism to release the token after
the page is unloaded. (However, the unload event is declared obsolete
and need not be supported by all browsers. Especially mobile browsers
are known not to issue it. On the other hand, this seems to be the only
mechanism usable for automatic token release.)

Buttons which cause the editor to close (including a new "save and exit"
button) release the token independently of the unload handler.

The "show PDF" button opens the PDF in a new tab, so it does not release
the token.

We also added logging to the JS console.