Allow more characters in task names (#31)
Currently, only lower case letters are allowed. However,
CMS allows [a-zA-Z0-9_-] characters in a tasks code name.
So extend the set of allowed characters, but note that the
hyphen is excluded because the task names appear in file
names that are passed in command line arguments and initial
hyphens might be misinterpreted as command line options.