autotranslate: two UI improvements (#36)
* autotranslate: put `row`s into a container element

There is currently horizontal scrolling in the machine translation window
because items are overflowing. Using the `container-fluid` class for the
container of the elements with the `row` class and removing the width and
padding eliminates the overflow; making `overflow-y: auto` unnecessary.

* autotranslate: disallow horizontal resizing of to-be-translated textarea

The user interface does not adapt well to horizontal resizing:
the textarea can potentially go under the translated text.