Add administration icon in the navigation bar
Also, when the admin logs in, the staff overview page is displayed
instead of automatically redirecting to the admin interface. This
makes it much easier to use the site as an admin. In particular,
the "view site" in admin interface now works, while the admin
interface is still easy to reach via the navigation bar.

Closes #54.