Clarify what is PrintedDocument.file_path
Some interpreted it as an URL, some as a relative path in the
filesystem. Once we started using MEDIA_ROOT systematically,
this double play broke down.

Now, the file_path is a path relative to the media directory,
which is sometimes prepended with MEDIA_ROOT and sometimes with MEDIA_URL.